근태보기 소숫점 1자리(형식: demical)

+++ HanbizaMVC/HanbizaMVC/ViewModel/TotalAttendence.cs
@@ -4,14 +4,14 @@ namespace HanbizaMVC.ViewModel
 {
     public class TotalAttendence
     {
-        public double 총근로 { get; set; }
-        public double 소정근로 { get; set; }
-        public double 근태조정 { get; set; }
-        public double 초과근로 { get; set; }
-        public double 연장근로 { get; set; }
-        public double 야간근로 { get; set; }
-        public double 휴일근로 { get; set; }
-        public double 휴일연장 { get; set; }
-        public double 휴일야간 { get; set; }
+        public decimal 총근로 { get; set; }
+        public decimal 소정근로 { get; set; }
+        public decimal 근태조정 { get; set; }
+        public decimal 초과근로 { get; set; }
+        public decimal 연장근로 { get; set; }
+        public decimal 야간근로 { get; set; }
+        public decimal 휴일근로 { get; set; }
+        public decimal 휴일연장 { get; set; }
+        public decimal 휴일야간 { get; set; }
     }
 }
